+++ app/src/main/kotlin/com/simplemobiletools/calendar/pro/extensions/DateTime.kt
@@ -2,4 +2,4 @@ package com.simplemobiletools.calendar.pro.extensions
 
 import org.joda.time.DateTime
 
-fun DateTime.seconds() = (millis / 1000).toInt()
+fun DateTime.seconds() = millis / 1000L
