+++ app/src/main/kotlin/com/simplemobiletools/calendar/pro/helpers/Constants.kt
@@ -143,4 +143,4 @@ const val DELETE_SELECTED_OCCURRENCE = 0
 const val DELETE_FUTURE_OCCURRENCES = 1
 const val DELETE_ALL_OCCURRENCES = 2
 
-fun getNowSeconds() = (System.currentTimeMillis() / 1000).toInt()
+fun getNowSeconds() = System.currentTimeMillis() / 1000L
